Create param abstract class, and convert ListVolumesParam to it
