+++ src/main/java/com/spotify/docker/client/DockerClient.java
@@ -2509,23 +2509,10 @@ public interface DockerClient extends Closeable {
    * Parameters for {@link #listVolumes(ListVolumesParam...)}.
    * @since Docker 1.9, API version 1.21
    */
-  class ListVolumesParam {
-    private final String name;
-    private final String value;
+  class ListVolumesParam extends Param {
 
     private ListVolumesParam(final String name, final String value) {
-      this.name = name;
-      this.value = value;
-    }
-
-    /**
-     * Parameter name.
-     *
-     * @return name of parameter
-     * @since Docker 1.9, API version 1.21
-     */
-    public String name() {
-      return name;
+      super(name, value);
     }
 
     /**
@@ -2536,16 +2523,6 @@ public interface DockerClient extends Closeable {
      */
     public static ListVolumesParam name(final String name) {
       return filter("name", name);
-    }
-
-    /**
-     * Parameter value.
-     *
-     * @return value of parameter
-     * @since Docker 1.9, API version 1.21
-     */
-    public String value() {
-      return value;
     }
 
     /**
@@ -2593,25 +2570,6 @@ public interface DockerClient extends Closeable {
       return filter("driver", driver);
     }
 
-    @Override
-    public boolean equals(final Object obj) {
-      if (this == obj) {
-        return true;
-      }
-      if (obj == null || getClass() != obj.getClass()) {
-        return false;
-      }
-
-      ListVolumesParam that = (ListVolumesParam) obj;
-
-      return Objects.equals(name, that.name)
-              && Objects.equals(value, that.value);
-    }
-
-    @Override
-    public int hashCode() {
-      return Objects.hash(name, value);
-    }
   }
 
   /**
